backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
pgvector>=0.2.0
alembic>=1.11.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
google-genai>=0.2.0
PyPDF2>=3.0.0
